Tried to rename everything to SerializedTransferProtocol
-- After some consideration and seeing just how much stuff needs to be renamed, I have decided to go with one more rename, but using something closer to the original: SerializedTransferProtocol.

Serial really implies "serial", a well-known wire-based data transmission protocols. Serialized, on the other hand, is technically how ANY data is transmitted using ANY protocol and this is more or less what this project does: Serializes different python or C/C++ objects, CRC and COBS encodes them and sends them over. For now it has only been tested with Serial connection, but I am, actively working on adding a custom ZeroMQ backbone that mimics how Serial is used for pySerial (and Stream for Microcontroller). The goal is to get ZeroMQ python class support and provide a separate C-dll that implements the necessary class methods from C to be dynamically linked to Unity projects (the final target of this project)