Implemented ZeroMQSerial class
-- Fully implemented and tested the runtime behavior of the ZeroMQSerial class.

ZeroMQSerial uses ZeroMQ communication protocol to establish a bidirectional socket-based connection with another ZeroMQ instance using the specified protocol, address and port. It is implemented to use very similar API to pySerial's Serial class and it can be used as a drop-in replacement for the Serial class by the SerializedTransferProtocol class.

This implementation allows using Serial to communicate with microcontrollers and ZeroMQSerial to communicate with other processes (cores) and machines (computers). This way, the library can 'talk' to more or less any other device as long as it can use Serial or ZeroMQ.

-- Also changed the error messages for COBS and CRC processors to include the static integer codes for specific errors. This is needed for the error-loging mechanism to be implemented in near future to re-route certain non-critical errors to log files (helpful for contexts like Uno where errors more or less are guaranteed to happen and it is only a question of 'when').